--- a/src/onefetch/info.rs
+++ b/src/onefetch/info.rs
@@ -6,7 +6,7 @@ use {
     colored::{Color, ColoredString, Colorize},
     git2::Repository,
     regex::Regex,
-    tokio::process::Command,
+    std::process::Command,
 };
 
 pub struct Info {
@@ -201,31 +201,19 @@ impl std::fmt::Display for Info {
 }
 
 impl Info {
-    #[tokio::main]
-    pub async fn new(config: Cli) -> Result<Info> {
+    pub fn new(config: Cli) -> Result<Info> {
         let repo = Repository::discover(&config.path)
             .chain_err(|| "Could not find a valid git repo on the current path")?;
         let workdir = repo.workdir().chain_err(|| "Unable to run onefetch on bare git repo")?;
         let workdir_str = workdir.to_str().unwrap();
         let (languages_stats, number_of_lines) =
             Language::get_language_statistics(workdir_str, &config.excluded)?;
-
-        let (
-            git_history,
-            (number_of_tags, number_of_branches),
-            (git_v, git_user),
-            version,
-            pending,
-            repo_size,
-        ) = futures::join!(
-            Info::get_git_history(workdir_str, config.no_merges),
-            Info::get_number_of_tags_branches(workdir_str),
-            Info::get_git_version_and_username(workdir_str),
-            Info::get_version(workdir_str),
-            Info::get_pending_changes(workdir_str),
-            Info::get_packed_size(workdir_str)
-        );
-
+        let git_history = Info::get_git_history(workdir_str, config.no_merges);
+        let (number_of_tags, number_of_branches) = Info::get_number_of_tags_branches(workdir_str);
+        let (git_v, git_user) = Info::get_git_version_and_username(workdir_str);
+        let version = Info::get_version(workdir_str);
+        let pending = Info::get_pending_changes(workdir_str);
+        let repo_size = Info::get_packed_size(workdir_str);
         let (repository_name, repository_url) = Info::get_repo_name_and_url(&repo);
         let current_commit_info = Info::get_current_commit_info(&repo);
         let creation_date = Info::get_creation_date(&git_history);
@@ -267,7 +255,7 @@ impl Info {
         })
     }
 
-    async fn get_git_history(dir: &str, no_merges: bool) -> Vec<String> {
+    fn get_git_history(dir: &str, no_merges: bool) -> Vec<String> {
         let mut args = vec!["-C", dir, "log"];
         if no_merges {
             args.push("--no-merges");
@@ -275,18 +263,17 @@ impl Info {
 
         args.push("--pretty=%cr\t%ae\t%an");
 
-        let output = Command::new("git").args(args).output().await.expect("Failed to execute git.");
+        let output = Command::new("git").args(args).output().expect("Failed to execute git.");
 
         let output = String::from_utf8_lossy(&output.stdout);
         output.lines().map(|x| x.to_string()).collect::<Vec<_>>()
     }
 
-    async fn get_number_of_tags_branches(dir: &str) -> (usize, usize) {
+    fn get_number_of_tags_branches(dir: &str) -> (usize, usize) {
         let tags = {
             let output = Command::new("git")
                 .args(vec!["-C", dir, "tag"])
                 .output()
-                .await
                 .expect("Failed to execute git.");
 
             let tags = String::from_utf8_lossy(&output.stdout);
@@ -298,7 +285,6 @@ impl Info {
             let output = Command::new("git")
                 .args(vec!["-C", dir, "branch", "-r"])
                 .output()
-                .await
                 .expect("Failed to execute git.");
 
             let branches = String::from_utf8_lossy(&output.stdout);
@@ -401,9 +387,9 @@ impl Info {
         authors
     }
 
-    async fn get_git_version_and_username(dir: &str) -> (String, String) {
+    fn get_git_version_and_username(dir: &str) -> (String, String) {
         let version =
-            Command::new("git").arg("--version").output().await.expect("Failed to execute git.");
+            Command::new("git").arg("--version").output().expect("Failed to execute git.");
         let version = String::from_utf8_lossy(&version.stdout).replace('\n', "");
 
         let username = Command::new("git")
@@ -413,13 +399,12 @@ impl Info {
             .arg("--get")
             .arg("user.name")
             .output()
-            .await
             .expect("Failed to execute git.");
         let username = String::from_utf8_lossy(&username.stdout).replace('\n', "");
         (version, username)
     }
 
-    async fn get_version(dir: &str) -> Result<String> {
+    fn get_version(dir: &str) -> Result<String> {
         let output = Command::new("git")
             .arg("-C")
             .arg(dir)
@@ -427,7 +412,6 @@ impl Info {
             .arg("--abbrev=0")
             .arg("--tags")
             .output()
-            .await
             .expect("Failed to execute git.");
 
         let output = String::from_utf8_lossy(&output.stdout);
@@ -444,14 +428,13 @@ impl Info {
         number_of_commits.to_string()
     }
 
-    async fn get_pending_changes(dir: &str) -> Result<String> {
+    fn get_pending_changes(dir: &str) -> Result<String> {
         let output = Command::new("git")
             .arg("-C")
             .arg(dir)
             .arg("status")
             .arg("--porcelain")
             .output()
-            .await
             .expect("Failed to execute git.");
 
         let output = String::from_utf8_lossy(&output.stdout);
@@ -493,14 +476,13 @@ impl Info {
         }
     }
 
-    async fn get_packed_size(dir: &str) -> Result<String> {
+    fn get_packed_size(dir: &str) -> Result<String> {
         let output = Command::new("git")
             .arg("-C")
             .arg(dir)
             .arg("count-objects")
             .arg("-vH")
             .output()
-            .await
             .expect("Failed to execute git.");
 
         let output = String::from_utf8_lossy(&output.stdout);
@@ -517,7 +499,6 @@ impl Info {
             .arg(dir)
             .arg("ls-files")
             .output()
-            .await
             .expect("Failed to execute git.");
         // To check if command executed successfully or not
         let error = &output.stderr;