## With async/await on the linux repo:

```bash
#Response times for async calls
o2sh ~/dev/onefetch (telemetry) > onefetch ../linux/
get_number_of_tags_branches --> 40.105318ms
get_git_version_and_username --> 39.971554ms
get_packed_size --> 89.731579ms
get_version --> 189.732919ms
get_pending_changes --> 199.959989ms
get_git_history --> 18.836888247s
```

`get_git_history` is overwhelmingly slower than the other functions;even with async/await the execution is still very synchronous-like.

```bash
time onefetch ../linux

real	0m22,442s
user	0m48,883s
sys	0m6,863s
```
## Without async/await on the linux repo:

```bash
time onefetch ../linux

real	0m15,631s
user	0m22,745s
sys	0m4,017s
```
-------
It may have been a false good idea to add tokio runtime - with its huge added complexity - to onefetch. Especially considering the use case - only one very slow function `get_git_history` -. Without async/await, onefetch appears to compile and run faster and use less cpu resources (cf. `user` time).